++ Add calculation history and improve display handling
- Implement history tracking for previous calculations
- Move parser implementation to separate source file
- Improve display formatting and zero handling
- Adjust UI layout to accommodate history display
