+++ includes/calculator.h
@@ -1,9 +1,11 @@
 #pragma once
 #include <string>
+#include <vector>
 
 struct CalculatorState {
     std::string display;
     std::string expression;
+    std::vector<std::string> history;
     double operand1;
     double operand2;
     char op;
